Adds tests for getNameAndValue
