+++ .changeset/tiny-trains-float.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/combobox': patch
+---
+
+Adds additional tests for internal Combobox utilities
